+++ packages/file-input/server/index.js
@@ -1,0 +1,1 @@
+export { default as uploadBuffer } from './uploadBuffer.js'