+++ packages/startupjs-ui/exports/_palette.tsx
@@ -1,0 +1,1 @@
+export { palette as default } from '@startupjs-ui/core'