+++ packages/startupjs-ui/server/index.js
@@ -1,0 +1,1 @@
+export { uploadBuffer } from '@startupjs-ui/file-input/server'