+++ packages/auto-suggest/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.16](https://github.com/startupjs/startupjs-ui/compare/v0.1.15...v0.1.16) (2026-02-10)
+
+**Note:** Version bump only for package @startupjs-ui/auto-suggest
+
+
+
+
+
+## [0.1.15](https://github.com/startupjs/startupjs-ui/compare/v0.1.14...v0.1.15) (2026-02-06)
+
+**Note:** Version bump only for package @startupjs-ui/auto-suggest
+
+
+
+
+
+## [0.1.14](https://github.com/startupjs/startupjs-ui/compare/v0.1.13...v0.1.14) (2026-02-06)
+
+
+### Features
+
+* add size for AutoSuggestProps ([#8](https://github.com/startupjs/startupjs-ui/issues/8)) ([c0cd5b0](https://github.com/startupjs/startupjs-ui/commit/c0cd5b0d84513e98c0e70cbcb0a5073db334c762))
+
+
+
+
+
+## [0.1.13](https://github.com/startupjs/startupjs-ui/compare/v0.1.12...v0.1.13) (2026-02-03)
+
+**Note:** Version bump only for package @startupjs-ui/auto-suggest
+
+
+
+
+
+## [0.1.12](https://github.com/startupjs/startupjs-ui/compare/v0.1.11...v0.1.12) (2026-01-21)
+
+**Note:** Version bump only for package @startupjs-ui/auto-suggest
+
+
+
+
+
+## [0.1.11](https://github.com/startupjs/startupjs-ui/compare/v0.1.10...v0.1.11) (2026-01-20)
+
+**Note:** Version bump only for package @startupjs-ui/auto-suggest
+
+
+
+
+
+## [0.1.10](https://github.com/startupjs/startupjs-ui/compare/v0.1.9...v0.1.10) (2026-01-18)
+
+**Note:** Version bump only for package @startupjs-ui/auto-suggest
+
+
+
+
+
 ## [0.1.5](https://github.com/startupjs/startupjs-ui/compare/v0.1.4...v0.1.5) (2025-12-29)
 
 **Note:** Version bump only for package @startupjs-ui/auto-suggest