+++ packages/core/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.11](https://github.com/startupjs/startupjs-ui/compare/v0.1.10...v0.1.11) (2026-01-20)
+
+
+### Bug Fixes
+
+* **core:** change shadow() stylus helper to have a constant black color for different levels ([5a49a24](https://github.com/startupjs/startupjs-ui/commit/5a49a24fcb52b147467fb528f4f432470f640794))
+
+
+
+
+
 ## [0.1.5](https://github.com/startupjs/startupjs-ui/compare/v0.1.4...v0.1.5) (2025-12-29)
 
 **Note:** Version bump only for package @startupjs-ui/core