+++ packages/object-input/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.16](https://github.com/startupjs/startupjs-ui/compare/v0.1.15...v0.1.16) (2026-02-10)
+
+**Note:** Version bump only for package @startupjs-ui/object-input
+
+
+
+
+
+## [0.1.13](https://github.com/startupjs/startupjs-ui/compare/v0.1.12...v0.1.13) (2026-02-03)
+
+**Note:** Version bump only for package @startupjs-ui/object-input
+
+
+
+
+
+## [0.1.12](https://github.com/startupjs/startupjs-ui/compare/v0.1.11...v0.1.12) (2026-01-21)
+
+**Note:** Version bump only for package @startupjs-ui/object-input
+
+
+
+
+
+## [0.1.11](https://github.com/startupjs/startupjs-ui/compare/v0.1.10...v0.1.11) (2026-01-20)
+
+**Note:** Version bump only for package @startupjs-ui/object-input
+
+
+
+
+
+## [0.1.10](https://github.com/startupjs/startupjs-ui/compare/v0.1.9...v0.1.10) (2026-01-18)
+
+**Note:** Version bump only for package @startupjs-ui/object-input
+
+
+
+
+
+## [0.1.9](https://github.com/startupjs/startupjs-ui/compare/v0.1.8...v0.1.9) (2026-01-16)
+
+**Note:** Version bump only for package @startupjs-ui/object-input
+
+
+
+
+
+## [0.1.8](https://github.com/startupjs/startupjs-ui/compare/v0.1.7...v0.1.8) (2026-01-08)
+
+**Note:** Version bump only for package @startupjs-ui/object-input
+
+
+
+
+
 ## [0.1.7](https://github.com/startupjs/startupjs-ui/compare/v0.1.6...v0.1.7) (2026-01-06)
 
 **Note:** Version bump only for package @startupjs-ui/object-input